requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2022.12.7
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.54.0
django-appconf==1.0.4
django-axes==5.21.0
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-extra-views==0.14.0
django-imagekit==4.0.2
django-inplaceedit==1.4.1
django-ipware==3.0.7
django-summernote==0.8.20.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
django-redis==5.2.0
dj-rest-auth==3.0.0
djoser==2.1.0
idna==3.2
mysqlclient==2.0.3
numpy
oauthlib==3.2.2
opencv-python
packaging==21.0
pilkit==2.0
Pillow==8.3.2
pycparser==2.20
PyJWT
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.4
urllib3==1.26.6
webencodings==0.5.1